% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPal.R
\name{createPal}
\alias{createPal}
\title{Create a palette of colors based on user input}
\usage{
createPal(user_cols, num_cols, full = TRUE, plot = TRUE)
}
\arguments{
\item{user_cols}{a vector of R default colors or a vector of HEX}

\item{num_cols}{number of colors wanted in the output}

\item{full}{TRUE/FALSE, whether the full gradient should be plotted or not}
}
\value{
a full gradient plot, your palette plot, vector of HEX codes
}
\description{
Create a palette of colors based on user input
}
